Synkron service bruker preprosesseringservice
