+++ src/main/kotlin/no/nav/helse/prosessering/v1/asynkron/JournalforingsStream.kt
@@ -3,7 +3,7 @@ package no.nav.helse.prosessering.v1.asynkron
 import no.nav.helse.CorrelationId
 import no.nav.helse.HttpError
 import no.nav.helse.aktoer.AktoerId
-import no.nav.helse.gosys.JoarkGateway
+import no.nav.helse.joark.JoarkGateway
 import no.nav.helse.kafka.KafkaConfig
 import no.nav.helse.kafka.PauseableKafkaStreams
 import no.nav.helse.prosessering.v1.PreprossesertMeldingV1
